# Conflicts:
#	backend/ModelLayer/Migrations/20200609191420_init.Designer.cs
#	backend/ModelLayer/Migrations/20200609191420_init.cs
#	backend/ModelLayer/Migrations/20200610204442_init.Designer.cs
#	backend/ModelLayer/Migrations/20200610204442_init.cs
#	backend/ModelLayer/Migrations/20200614115428_init.Designer.cs
#	backend/ModelLayer/Migrations/20200614115428_init.cs
#	backend/ModelLayer/Migrations/CrmContextModelSnapshot.cs
#	backend/WebApi/Startup.cs
#	frontend/src/app/organizations/organizations-list/organizations-list.component.ts
#	frontend/src/app/organizations/organizations.module.ts

+++ frontend/src/app/events/events-list/events-list.component.ts
@@ -7,6 +7,7 @@ import { EventsAddComponent } from '../events-add/events-add.component';
 import { EventsDetailComponent } from '../events-detail/events-detail.component';
 import { MatSort } from '@angular/material/sort';
 import { EventsInfoComponent } from '../events-info/events-info.component';
+import { DeleteEntryDialogComponent } from '../../shared/form/delete-entry-dialog/delete-entry-dialog.component';
 
 export class EventDtoGroup implements EventDto {
   id: number;
@@ -37,7 +38,7 @@ export class EventsListComponent implements OnInit {
   constructor(
     private service: EventService,
     private dialog: MatDialog) {
-   }
+  }
 
   ngOnInit() {
     this.init();
@@ -64,13 +65,13 @@ export class EventsListComponent implements OnInit {
   }
 
   addContact() {
-    const dialogRef = this.dialog.open(EventsAddComponent);
+    const dialogRef = this.dialog.open(EventsAddComponent, { disableClose: true });
     dialogRef.afterClosed().subscribe(x => this.init());
   }
 
   callEdit(id: number) {
     this.service.getById(id).subscribe(x => {
-      const dialogRef = this.dialog.open(EventsDetailComponent, { data: x });
+      const dialogRef = this.dialog.open(EventsDetailComponent, { data: x, disableClose: true });
       dialogRef.afterClosed().subscribe(y => this.init());
     });
   }
@@ -85,7 +86,16 @@ export class EventsListComponent implements OnInit {
   }
 
   deleteEvent(id: number) {
-    this.service.delete(id).subscribe(x => this.init());
+    const deleteDialogRef = this.dialog.open(DeleteEntryDialogComponent, {
+      data: 'Event',
+      disableClose: true
+    });
+
+    deleteDialogRef.afterClosed().subscribe((deleteResult) => {
+      if (deleteResult.delete) {
+        this.service.delete(id).subscribe(x => this.init());
+      }
+    });
   }
 
   toggleSelection() {
@@ -103,21 +113,9 @@ export class EventsListComponent implements OnInit {
       eventsFiltered = events;
     }
     eventsFiltered.forEach(x => {
-        const week = this.getWeekNumber(new Date(x.date));
-        if (week !== this.weekNumber) {
-          this.weekNumber = week;
-          this.dataSource.push({
-            date: x.date,
-            duration: x.duration,
-            id: x.id,
-            time: x.time,
-            contacts: x.contacts,
-            name: x.name,
-            participated: x.participated,
-            weekNumber: week,
-            isGroupBy: true
-          });
-        }
+      const week = this.getWeekNumber(new Date(x.date));
+      if (week !== this.weekNumber) {
+        this.weekNumber = week;
         this.dataSource.push({
           date: x.date,
           duration: x.duration,
@@ -126,9 +124,21 @@ export class EventsListComponent implements OnInit {
           contacts: x.contacts,
           name: x.name,
           participated: x.participated,
-          weekNumber: 0,
-          isGroupBy: false
+          weekNumber: week,
+          isGroupBy: true
         });
+      }
+      this.dataSource.push({
+        date: x.date,
+        duration: x.duration,
+        id: x.id,
+        time: x.time,
+        contacts: x.contacts,
+        name: x.name,
+        participated: x.participated,
+        weekNumber: 0,
+        isGroupBy: false
+      });
     });
   }
 
@@ -143,7 +153,7 @@ export class EventsListComponent implements OnInit {
 
   openInfo(id: number) {
     this.service.getById(id).subscribe(x => {
-      this.dialog.open(EventsInfoComponent, { data: x });
+      this.dialog.open(EventsInfoComponent, { data: x, disableClose: true });
     });
   }
